Refactor: move code to compute the tx hashes and extract amounts out of sign_psbt.
Created separate utility modules instead.
Deleted the 'update_hashes_with_map_value' module, as it was only used in
the new txhashes.c module.

Cleaned up interface of the various compute_sighash_* functions.
Avoid taking the key placeholder as an argument, and declare some
of the arguments as const pointers.
