+++ src/handler/sign_psbt/txhashes.c
@@ -1,0 +1,694 @@
+#include "txhashes.h"
+#include "amount_from_psbt.h"
+
+#include "../../common/psbt.h"
+#include "../../error_codes.h"
+
+#include "../lib/get_merkleized_map.h"
+#include "../lib/get_merkleized_map_value.h"
+#include "../lib/stream_merkleized_map_value.h"
+
+/* BIP0341 tags for computing the tagged hashes when computing he sighash */
+static const uint8_t BIP0341_sighash_tag[] = {'T', 'a', 'p', 'S', 'i', 'g', 'h', 'a', 's', 'h'};
+
+typedef struct {
+    cx_hash_t *hash_unprefixed;
+    cx_hash_t *hash_prefixed;
+} callback_state_t;
+
+static void cb_process_len(size_t len, void *cb_state) {
+    callback_state_t *state = (callback_state_t *) cb_state;
+
+    if (state->hash_prefixed != NULL) {
+        crypto_hash_update_varint(state->hash_prefixed, len);
+    }
+}
+
+static void cb_process_data(buffer_t *data, void *cb_state) {
+    callback_state_t *state = (callback_state_t *) cb_state;
+
+    size_t data_len = data->size - data->offset;
+    uint8_t *data_start_ptr = data->ptr + data->offset;
+
+    if (state->hash_prefixed != NULL) {
+        crypto_hash_update(state->hash_prefixed, data_start_ptr, data_len);
+    }
+    if (state->hash_unprefixed != NULL) {
+        crypto_hash_update(state->hash_unprefixed, data_start_ptr, data_len);
+    }
+}
+
+static int update_hashes_with_map_value(dispatcher_context_t *dispatcher_context,
+                                        const merkleized_map_commitment_t *map,
+                                        const uint8_t *key,
+                                        int key_len,
+                                        cx_hash_t *hash_unprefixed,
+                                        cx_hash_t *hash_prefixed) {
+    LOG_PROCESSOR(__FILE__, __LINE__, __func__);
+
+    callback_state_t cb_state = {.hash_unprefixed = hash_unprefixed,
+                                 .hash_prefixed = hash_prefixed};
+
+    return call_stream_merkleized_map_value(dispatcher_context,
+                                            map,
+                                            key,
+                                            key_len,
+                                            cb_process_len,
+                                            cb_process_data,
+                                            &cb_state);
+}
+
+// Updates the hash_context with the output of given index
+// returns -1 on error. 0 on success.
+static int hash_output_n(dispatcher_context_t *dc,
+                         const sign_psbt_state_t *st,
+                         cx_hash_t *hash_context,
+                         unsigned int index) {
+    if (index >= st->n_outputs) {
+        return -1;
+    }
+
+    // get this output's map
+    merkleized_map_commitment_t ith_map;
+
+    int res = call_get_merkleized_map(dc, st->outputs_root, st->n_outputs, index, &ith_map);
+    if (res < 0) {
+        return -1;
+    }
+
+    // get output's amount
+    uint8_t amount_raw[8];
+    if (8 != call_get_merkleized_map_value(dc,
+                                           &ith_map,
+                                           (uint8_t[]){PSBT_OUT_AMOUNT},
+                                           1,
+                                           amount_raw,
+                                           8)) {
+        return -1;
+    }
+
+    crypto_hash_update(hash_context, amount_raw, 8);
+
+    // get output's scriptPubKey
+
+    uint8_t out_script[MAX_OUTPUT_SCRIPTPUBKEY_LEN];
+    int out_script_len = call_get_merkleized_map_value(dc,
+                                                       &ith_map,
+                                                       (uint8_t[]){PSBT_OUT_SCRIPT},
+                                                       1,
+                                                       out_script,
+                                                       sizeof(out_script));
+    if (out_script_len == -1) {
+        return -1;
+    }
+
+    crypto_hash_update_varint(hash_context, out_script_len);
+    crypto_hash_update(hash_context, out_script, out_script_len);
+    return 0;
+}
+
+// Updates the hash_context with the network serialization of all the outputs
+// returns -1 on error. 0 on success.
+static int hash_outputs(dispatcher_context_t *dc,
+                        const sign_psbt_state_t *st,
+                        cx_hash_t *hash_context) {
+    for (unsigned int i = 0; i < st->n_outputs; i++) {
+        if (hash_output_n(dc, st, hash_context, i)) {
+            return -1;
+        }
+    }
+    return 0;
+}
+
+bool __attribute__((noinline))
+compute_tx_hashes(dispatcher_context_t *dc, sign_psbt_state_t *st, tx_hashes_t *hashes) {
+    {
+        // compute sha_prevouts and sha_sequences
+        cx_sha256_t sha_prevouts_context, sha_sequences_context;
+
+        // compute hashPrevouts and hashSequence
+        cx_sha256_init(&sha_prevouts_context);
+        cx_sha256_init(&sha_sequences_context);
+
+        for (unsigned int i = 0; i < st->n_inputs; i++) {
+            // get this input's map
+            merkleized_map_commitment_t ith_map;
+
+            int res = call_get_merkleized_map(dc, st->inputs_root, st->n_inputs, i, &ith_map);
+            if (res < 0) {
+                SEND_SW(dc, SW_INCORRECT_DATA);
+                return false;
+            }
+
+            // get prevout hash and output index for the i-th input
+            uint8_t ith_prevout_hash[32];
+            if (32 != call_get_merkleized_map_value(dc,
+                                                    &ith_map,
+                                                    (uint8_t[]){PSBT_IN_PREVIOUS_TXID},
+                                                    1,
+                                                    ith_prevout_hash,
+                                                    32)) {
+                SEND_SW(dc, SW_INCORRECT_DATA);
+                return false;
+            }
+
+            crypto_hash_update(&sha_prevouts_context.header, ith_prevout_hash, 32);
+
+            uint8_t ith_prevout_n_raw[4];
+            if (4 != call_get_merkleized_map_value(dc,
+                                                   &ith_map,
+                                                   (uint8_t[]){PSBT_IN_OUTPUT_INDEX},
+                                                   1,
+                                                   ith_prevout_n_raw,
+                                                   4)) {
+                SEND_SW(dc, SW_INCORRECT_DATA);
+                return false;
+            }
+
+            crypto_hash_update(&sha_prevouts_context.header, ith_prevout_n_raw, 4);
+
+            uint8_t ith_nSequence_raw[4];
+            if (4 != call_get_merkleized_map_value(dc,
+                                                   &ith_map,
+                                                   (uint8_t[]){PSBT_IN_SEQUENCE},
+                                                   1,
+                                                   ith_nSequence_raw,
+                                                   4)) {
+                // if no PSBT_IN_SEQUENCE is present, we must assume nSequence 0xFFFFFFFF
+                memset(ith_nSequence_raw, 0xFF, 4);
+            }
+
+            crypto_hash_update(&sha_sequences_context.header, ith_nSequence_raw, 4);
+        }
+
+        crypto_hash_digest(&sha_prevouts_context.header, hashes->sha_prevouts, 32);
+        crypto_hash_digest(&sha_sequences_context.header, hashes->sha_sequences, 32);
+    }
+
+    {
+        // compute sha_outputs
+        cx_sha256_t sha_outputs_context;
+        cx_sha256_init(&sha_outputs_context);
+
+        if (hash_outputs(dc, st, &sha_outputs_context.header) == -1) {
+            SEND_SW(dc, SW_INCORRECT_DATA);
+            return false;
+        }
+
+        crypto_hash_digest(&sha_outputs_context.header, hashes->sha_outputs, 32);
+    }
+
+    {
+        // compute sha_amounts and sha_scriptpubkeys
+        // TODO: could be skipped if there are no segwitv1 inputs to sign
+
+        cx_sha256_t sha_amounts_context, sha_scriptpubkeys_context;
+
+        cx_sha256_init(&sha_amounts_context);
+        cx_sha256_init(&sha_scriptpubkeys_context);
+
+        for (unsigned int i = 0; i < st->n_inputs; i++) {
+            // get this input's map
+            merkleized_map_commitment_t ith_map;
+
+            int res = call_get_merkleized_map(dc, st->inputs_root, st->n_inputs, i, &ith_map);
+            if (res < 0) {
+                SEND_SW(dc, SW_INCORRECT_DATA);
+                return false;
+            }
+
+            uint64_t in_amount;
+            uint8_t in_scriptPubKey[MAX_PREVOUT_SCRIPTPUBKEY_LEN];
+            size_t in_scriptPubKey_len;
+
+            if (0 > get_amount_scriptpubkey_from_psbt(dc,
+                                                      &ith_map,
+                                                      &in_amount,
+                                                      in_scriptPubKey,
+                                                      &in_scriptPubKey_len)) {
+                SEND_SW(dc, SW_INCORRECT_DATA);
+                return false;
+            }
+
+            uint8_t in_amount_le[8];
+            write_u64_le(in_amount_le, 0, in_amount);
+            crypto_hash_update(&sha_amounts_context.header, in_amount_le, 8);
+
+            crypto_hash_update_varint(&sha_scriptpubkeys_context.header, in_scriptPubKey_len);
+            crypto_hash_update(&sha_scriptpubkeys_context.header,
+                               in_scriptPubKey,
+                               in_scriptPubKey_len);
+        }
+
+        crypto_hash_digest(&sha_amounts_context.header, hashes->sha_amounts, 32);
+        crypto_hash_digest(&sha_scriptpubkeys_context.header, hashes->sha_scriptpubkeys, 32);
+    }
+
+    return true;
+}
+
+bool __attribute__((noinline)) compute_sighash_legacy(dispatcher_context_t *dc,
+                                                      const sign_psbt_state_t *st,
+                                                      const input_info_t *input,
+                                                      unsigned int input_index,
+                                                      uint8_t sighash[static 32]) {
+    LOG_PROCESSOR(__FILE__, __LINE__, __func__);
+
+    cx_sha256_t sighash_context;
+    cx_sha256_init(&sighash_context);
+
+    uint8_t tmp[4];
+    write_u32_le(tmp, 0, st->tx_version);
+    crypto_hash_update(&sighash_context.header, tmp, 4);
+
+    crypto_hash_update_varint(&sighash_context.header, st->n_inputs);
+
+    for (unsigned int i = 0; i < st->n_inputs; i++) {
+        // get this input's map
+        merkleized_map_commitment_t ith_map;
+
+        if (i != input_index) {
+            int res = call_get_merkleized_map(dc, st->inputs_root, st->n_inputs, i, &ith_map);
+            if (res < 0) {
+                SEND_SW(dc, SW_INCORRECT_DATA);
+                return false;
+            }
+        } else {
+            // Avoid requesting the same map unnecessarily
+            // (might be removed once a caching mechanism is implemented)
+            memcpy(&ith_map, &input->in_out.map, sizeof(input->in_out.map));
+        }
+
+        // get prevout hash and output index for the i-th input
+        uint8_t ith_prevout_hash[32];
+        if (32 != call_get_merkleized_map_value(dc,
+                                                &ith_map,
+                                                (uint8_t[]){PSBT_IN_PREVIOUS_TXID},
+                                                1,
+                                                ith_prevout_hash,
+                                                32)) {
+            SEND_SW(dc, SW_INCORRECT_DATA);
+            return false;
+        }
+
+        crypto_hash_update(&sighash_context.header, ith_prevout_hash, 32);
+
+        uint8_t ith_prevout_n_raw[4];
+        if (4 != call_get_merkleized_map_value(dc,
+                                               &ith_map,
+                                               (uint8_t[]){PSBT_IN_OUTPUT_INDEX},
+                                               1,
+                                               ith_prevout_n_raw,
+                                               4)) {
+            SEND_SW(dc, SW_INCORRECT_DATA);
+            return false;
+        }
+
+        crypto_hash_update(&sighash_context.header, ith_prevout_n_raw, 4);
+
+        if (i != input_index) {
+            // empty scriptcode
+            crypto_hash_update_u8(&sighash_context.header, 0x00);
+        } else {
+            if (!input->has_redeemScript) {
+                // P2PKH, the script_code is the prevout's scriptPubKey
+                crypto_hash_update_varint(&sighash_context.header, input->in_out.scriptPubKey_len);
+                crypto_hash_update(&sighash_context.header,
+                                   input->in_out.scriptPubKey,
+                                   input->in_out.scriptPubKey_len);
+            } else {
+                // P2SH, the script_code is the redeemScript
+
+                // update sighash_context with the length-prefixed redeem script
+                int redeemScript_len =
+                    update_hashes_with_map_value(dc,
+                                                 &input->in_out.map,
+                                                 (uint8_t[]){PSBT_IN_REDEEM_SCRIPT},
+                                                 1,
+                                                 NULL,
+                                                 &sighash_context.header);
+
+                if (redeemScript_len < 0) {
+                    PRINTF("Error fetching redeemScript\n");
+                    SEND_SW(dc, SW_INCORRECT_DATA);
+                    return false;
+                }
+            }
+        }
+
+        uint8_t ith_nSequence_raw[4];
+        if (4 != call_get_merkleized_map_value(dc,
+                                               &ith_map,
+                                               (uint8_t[]){PSBT_IN_SEQUENCE},
+                                               1,
+                                               ith_nSequence_raw,
+                                               4)) {
+            // if no PSBT_IN_SEQUENCE is present, we must assume nSequence 0xFFFFFFFF
+            memset(ith_nSequence_raw, 0xFF, 4);
+        }
+
+        crypto_hash_update(&sighash_context.header, ith_nSequence_raw, 4);
+    }
+
+    // outputs
+    crypto_hash_update_varint(&sighash_context.header, st->n_outputs);
+    if (hash_outputs(dc, st, &sighash_context.header) == -1) {
+        SEND_SW(dc, SW_INCORRECT_DATA);
+        return false;
+    }
+
+    // nLocktime
+    write_u32_le(tmp, 0, st->locktime);
+    crypto_hash_update(&sighash_context.header, tmp, 4);
+
+    // hash type
+    write_u32_le(tmp, 0, input->sighash_type);
+    crypto_hash_update(&sighash_context.header, tmp, 4);
+
+    // compute sighash
+    crypto_hash_digest(&sighash_context.header, sighash, 32);
+    cx_hash_sha256(sighash, 32, sighash, 32);
+
+    return true;
+}
+
+bool __attribute__((noinline)) compute_sighash_segwitv0(dispatcher_context_t *dc,
+                                                        const sign_psbt_state_t *st,
+                                                        const tx_hashes_t *hashes,
+                                                        const input_info_t *input,
+                                                        unsigned int input_index,
+                                                        uint8_t sighash[static 32]) {
+    LOG_PROCESSOR(__FILE__, __LINE__, __func__);
+
+    cx_sha256_t sighash_context;
+    cx_sha256_init(&sighash_context);
+
+    uint8_t tmp[8];
+    uint8_t sighash_byte = (uint8_t) (input->sighash_type & 0xFF);
+
+    // nVersion
+    write_u32_le(tmp, 0, st->tx_version);
+    crypto_hash_update(&sighash_context.header, tmp, 4);
+
+    {
+        uint8_t dbl_hash[32];
+
+        memset(dbl_hash, 0, 32);
+        // add to hash: hashPrevouts = sha256(sha_prevouts)
+        if (!(sighash_byte & SIGHASH_ANYONECANPAY)) {
+            cx_hash_sha256(hashes->sha_prevouts, 32, dbl_hash, 32);
+        }
+
+        crypto_hash_update(&sighash_context.header, dbl_hash, 32);
+
+        memset(dbl_hash, 0, 32);
+        // add to hash: hashSequence sha256(sha_sequences)
+        if (!(sighash_byte & SIGHASH_ANYONECANPAY) && (sighash_byte & 0x1f) != SIGHASH_SINGLE &&
+            (sighash_byte & 0x1f) != SIGHASH_NONE) {
+            cx_hash_sha256(hashes->sha_sequences, 32, dbl_hash, 32);
+        }
+        crypto_hash_update(&sighash_context.header, dbl_hash, 32);
+    }
+
+    {
+        // outpoint (32-byte prevout hash, 4-byte index)
+
+        // get prevout hash and output index for the current input
+        uint8_t prevout_hash[32];
+        if (32 != call_get_merkleized_map_value(dc,
+                                                &input->in_out.map,
+                                                (uint8_t[]){PSBT_IN_PREVIOUS_TXID},
+                                                1,
+                                                prevout_hash,
+                                                32)) {
+            SEND_SW(dc, SW_INCORRECT_DATA);
+            return false;
+        }
+
+        crypto_hash_update(&sighash_context.header, prevout_hash, 32);
+
+        uint8_t prevout_n_raw[4];
+        if (4 != call_get_merkleized_map_value(dc,
+                                               &input->in_out.map,
+                                               (uint8_t[]){PSBT_IN_OUTPUT_INDEX},
+                                               1,
+                                               prevout_n_raw,
+                                               4)) {
+            SEND_SW(dc, SW_INCORRECT_DATA);
+            return false;
+        }
+
+        crypto_hash_update(&sighash_context.header, prevout_n_raw, 4);
+    }
+
+    // scriptCode
+    if (is_p2wpkh(input->script, input->script_len)) {
+        // P2WPKH(script[2:22])
+        crypto_hash_update_u32(&sighash_context.header, 0x1976a914);
+        crypto_hash_update(&sighash_context.header, input->script + 2, 20);
+        crypto_hash_update_u16(&sighash_context.header, 0x88ac);
+    } else if (is_p2wsh(input->script, input->script_len)) {
+        // P2WSH
+
+        // update sighash_context.header with the length-prefixed witnessScript,
+        // and also compute sha256(witnessScript)
+        cx_sha256_t witnessScript_hash_context;
+        cx_sha256_init(&witnessScript_hash_context);
+
+        int witnessScript_len = update_hashes_with_map_value(dc,
+                                                             &input->in_out.map,
+                                                             (uint8_t[]){PSBT_IN_WITNESS_SCRIPT},
+                                                             1,
+                                                             &witnessScript_hash_context.header,
+                                                             &sighash_context.header);
+
+        if (witnessScript_len < 0) {
+            PRINTF("Error fetching witnessScript\n");
+            SEND_SW(dc, SW_INCORRECT_DATA);
+            return false;
+        }
+
+        uint8_t witnessScript_hash[32];
+        crypto_hash_digest(&witnessScript_hash_context.header, witnessScript_hash, 32);
+
+        // check that script == P2WSH(witnessScript)
+        if (input->script_len != 2 + 32 || input->script[0] != 0x00 || input->script[1] != 0x20 ||
+            memcmp(input->script + 2, witnessScript_hash, 32) != 0) {
+            PRINTF("Mismatching witnessScript\n");
+
+            SEND_SW_EC(dc, SW_INCORRECT_DATA, EC_SIGN_PSBT_MISMATCHING_WITNESS_SCRIPT);
+            return false;
+        }
+    } else {
+        PRINTF("Invalid or unsupported script in segwit transaction\n");
+        SEND_SW(dc, SW_INCORRECT_DATA);
+        return false;
+    }
+
+    {
+        // input value, taken from the WITNESS_UTXO field
+        uint8_t witness_utxo[8 + 1 + MAX_PREVOUT_SCRIPTPUBKEY_LEN];
+
+        int witness_utxo_len = call_get_merkleized_map_value(dc,
+                                                             &input->in_out.map,
+                                                             (uint8_t[]){PSBT_IN_WITNESS_UTXO},
+                                                             1,
+                                                             witness_utxo,
+                                                             sizeof(witness_utxo));
+        if (witness_utxo_len < 8) {
+            SEND_SW(dc, SW_INCORRECT_DATA);
+            return false;
+        }
+
+        crypto_hash_update(&sighash_context.header,
+                           witness_utxo,
+                           8);  // only the first 8 bytes (amount)
+    }
+
+    // nSequence
+    {
+        uint8_t nSequence_raw[4];
+        if (4 != call_get_merkleized_map_value(dc,
+                                               &input->in_out.map,
+                                               (uint8_t[]){PSBT_IN_SEQUENCE},
+                                               1,
+                                               nSequence_raw,
+                                               4)) {
+            // if no PSBT_IN_SEQUENCE is present, we must assume nSequence 0xFFFFFFFF
+            memset(nSequence_raw, 0xFF, 4);
+        }
+        crypto_hash_update(&sighash_context.header, nSequence_raw, 4);
+    }
+
+    {
+        // compute hashOutputs = sha256(sha_outputs)
+
+        uint8_t hashOutputs[32];
+        memset(hashOutputs, 0, 32);
+
+        if ((sighash_byte & 0x1f) != SIGHASH_SINGLE && (sighash_byte & 0x1f) != SIGHASH_NONE) {
+            cx_hash_sha256(hashes->sha_outputs, 32, hashOutputs, 32);
+
+        } else if ((sighash_byte & 0x1f) == SIGHASH_SINGLE && input_index < st->n_outputs) {
+            cx_sha256_t sha_output_context;
+            cx_sha256_init(&sha_output_context);
+            if (hash_output_n(dc, st, &sha_output_context.header, input_index) == -1) {
+                SEND_SW(dc, SW_INCORRECT_DATA);
+                return false;
+            }
+            crypto_hash_digest(&sha_output_context.header, hashOutputs, 32);
+            cx_hash_sha256(hashOutputs, 32, hashOutputs, 32);
+        }
+        crypto_hash_update(&sighash_context.header, hashOutputs, 32);
+    }
+
+    // nLocktime
+    write_u32_le(tmp, 0, st->locktime);
+    crypto_hash_update(&sighash_context.header, tmp, 4);
+
+    // sighash type
+    write_u32_le(tmp, 0, input->sighash_type);
+    crypto_hash_update(&sighash_context.header, tmp, 4);
+
+    // compute sighash
+    crypto_hash_digest(&sighash_context.header, sighash, 32);
+    cx_hash_sha256(sighash, 32, sighash, 32);
+
+    return true;
+}
+
+bool __attribute__((noinline)) compute_sighash_segwitv1(dispatcher_context_t *dc,
+                                                        const sign_psbt_state_t *st,
+                                                        const tx_hashes_t *hashes,
+                                                        const input_info_t *input,
+                                                        unsigned int input_index,
+                                                        const uint8_t *tapleaf_hash,
+                                                        uint8_t sighash[static 32]) {
+    LOG_PROCESSOR(__FILE__, __LINE__, __func__);
+
+    cx_sha256_t sighash_context;
+    crypto_tr_tagged_hash_init(&sighash_context, BIP0341_sighash_tag, sizeof(BIP0341_sighash_tag));
+    // the first 0x00 byte is not part of SigMsg
+    crypto_hash_update_u8(&sighash_context.header, 0x00);
+
+    uint8_t tmp[MAX(32, 8 + 1 + MAX_PREVOUT_SCRIPTPUBKEY_LEN)];
+
+    // hash type
+    uint8_t sighash_byte = (uint8_t) (input->sighash_type & 0xFF);
+    crypto_hash_update_u8(&sighash_context.header, sighash_byte);
+
+    // nVersion
+    write_u32_le(tmp, 0, st->tx_version);
+    crypto_hash_update(&sighash_context.header, tmp, 4);
+
+    // nLocktime
+    write_u32_le(tmp, 0, st->locktime);
+    crypto_hash_update(&sighash_context.header, tmp, 4);
+
+    if ((sighash_byte & 0x80) != SIGHASH_ANYONECANPAY) {
+        crypto_hash_update(&sighash_context.header, hashes->sha_prevouts, 32);
+        crypto_hash_update(&sighash_context.header, hashes->sha_amounts, 32);
+        crypto_hash_update(&sighash_context.header, hashes->sha_scriptpubkeys, 32);
+        crypto_hash_update(&sighash_context.header, hashes->sha_sequences, 32);
+    }
+
+    if ((sighash_byte & 3) != SIGHASH_NONE && (sighash_byte & 3) != SIGHASH_SINGLE) {
+        crypto_hash_update(&sighash_context.header, hashes->sha_outputs, 32);
+    }
+
+    // ext_flag
+    uint8_t ext_flag = tapleaf_hash != NULL ? 1 : 0;
+    // annex is not supported
+    const uint8_t annex_present = 0;
+    uint8_t spend_type = ext_flag * 2 + annex_present;
+    crypto_hash_update_u8(&sighash_context.header, spend_type);
+
+    if ((sighash_byte & 0x80) == SIGHASH_ANYONECANPAY) {
+        // outpoint (hash)
+        if (32 != call_get_merkleized_map_value(dc,
+                                                &input->in_out.map,
+                                                (uint8_t[]){PSBT_IN_PREVIOUS_TXID},
+                                                1,
+                                                tmp,
+                                                32)) {
+            SEND_SW(dc, SW_INCORRECT_DATA);
+            return false;
+        }
+        crypto_hash_update(&sighash_context.header, tmp, 32);
+
+        // outpoint (output index)
+        if (4 != call_get_merkleized_map_value(dc,
+                                               &input->in_out.map,
+                                               (uint8_t[]){PSBT_IN_OUTPUT_INDEX},
+                                               1,
+                                               tmp,
+                                               4)) {
+            SEND_SW(dc, SW_INCORRECT_DATA);
+            return false;
+        }
+        crypto_hash_update(&sighash_context.header, tmp, 4);
+
+        if (8 > call_get_merkleized_map_value(dc,
+                                              &input->in_out.map,
+                                              (uint8_t[]){PSBT_IN_WITNESS_UTXO},
+                                              1,
+                                              tmp,
+                                              8 + 1 + MAX_PREVOUT_SCRIPTPUBKEY_LEN)) {
+            SEND_SW(dc, SW_INCORRECT_DATA);
+            return false;
+        }
+
+        // amount
+        crypto_hash_update(&sighash_context.header, tmp, 8);
+
+        // scriptPubKey
+        crypto_hash_update_varint(&sighash_context.header, input->in_out.scriptPubKey_len);
+
+        crypto_hash_update(&sighash_context.header,
+                           input->in_out.scriptPubKey,
+                           input->in_out.scriptPubKey_len);
+
+        // nSequence
+        if (4 != call_get_merkleized_map_value(dc,
+                                               &input->in_out.map,
+                                               (uint8_t[]){PSBT_IN_SEQUENCE},
+                                               1,
+                                               tmp,
+                                               4)) {
+            // if no PSBT_IN_SEQUENCE is present, we must assume nSequence 0xFFFFFFFF
+            memset(tmp, 0xFF, 4);
+        }
+        crypto_hash_update(&sighash_context.header, tmp, 4);
+    } else {
+        // input_index
+        write_u32_le(tmp, 0, input_index);
+        crypto_hash_update(&sighash_context.header, tmp, 4);
+    }
+
+    // no annex
+
+    if ((sighash_byte & 3) == SIGHASH_SINGLE) {
+        // compute sha_output
+        cx_sha256_t sha_output_context;
+        cx_sha256_init(&sha_output_context);
+
+        if (hash_output_n(dc, st, &sha_output_context.header, input_index) == -1) {
+            SEND_SW(dc, SW_INCORRECT_DATA);
+            return false;
+        }
+        crypto_hash_digest(&sha_output_context.header, tmp, 32);
+
+        crypto_hash_update(&sighash_context.header, tmp, 32);
+    }
+
+    if (tapleaf_hash != NULL) {
+        // If spending a tapscript, append the Common Signature Message Extension per BIP-0342
+        crypto_hash_update(&sighash_context.header, tapleaf_hash, 32);
+        crypto_hash_update_u8(&sighash_context.header, 0x00);         // key_version
+        crypto_hash_update_u32(&sighash_context.header, 0xffffffff);  // no OP_CODESEPARATOR
+    }
+
+    crypto_hash_digest(&sighash_context.header, sighash, 32);
+
+    return true;
+}
